+++ aws-infrastructure/src/main/java/io/brau/AwsInfrastructureApp.java
@@ -24,7 +24,11 @@ public class AwsInfrastructureApp {
         SIHRepositoryStack repoStack = new SIHRepositoryStack(app, "SIHRepositoryStack", stackProps);
         DatalakeInfrastructureStack datalakeStack = new DatalakeInfrastructureStack(app, "DatalakeInfrastructureStack", stackProps, repoStack.getRepository());
         new DatabaseStack(app, "DatabaseStack", stackProps, datalakeStack.getVpc());
+        DatabasePublicStack databasePublicStack = new DatabasePublicStack(app, "DatabasePublicStack", stackProps, datalakeStack.getVpc());
         new ETLGlueStack(app, "ETLGlueStack", stackProps, datalakeStack.getBucket());
+        new ETLGlueJobStack(app, "ETLGlueJobStack", stackProps, datalakeStack.getBucket(),
+                datalakeStack.getVpc(), databasePublicStack.getInstance(), databasePublicStack.getSecretArn(),
+                databasePublicStack.getGlueSecurityGroup());
 
         app.synth();
     }